refactor(sync-core): tidy cross-tab status wire, env detection, and message types
Three small decoupling passes on the cross-tab modules:

- Centralize the leader-status wire format in a new `leaderStatus.ts`
  (`statusEventToMessage` / `messageToStatusEvent` / `toStatusChangeEvent`),
  so the event<->message mapping and its error-reason handling live in one
  place instead of being duplicated across the router and follower receiver.
- Move `BroadcastChannel` / `navigator.locks` detection out of the
  orchestrator into `defaultEnvironment.ts` (renamed from
  `defaultBrowserContext.ts`), grouping all browser-global resolution with
  the default browser context.
- Group `CrossTabMessage` into per-concern sub-unions (`StatusMessage`,
  `TransportMessage`, `PresenceMessage`, `LeadershipMessage`) so it's clear
  which module owns which arms.
